requirements.txt
opencv_python==4.1.1.26
PyYAML==5.1.2
pyinstaller==3.5
cython==0.29.14